homework9

+++ app/src/main/java/com/example/app/domain/model/TopicModel.kt
@@ -1,6 +1,6 @@
 package com.example.app.domain.model
 
 data class TopicModel(
-    val maxId: Int,
+    val maxId: Long,
     val name: String
 )
